add dynamic instruction compiler
Instead of interpreting each instruction, now the first time a given
instruction value is encountered a dynamic Java class will be generated
and used. A simple infinite-loop test measures an ~8.7x speedup.

There are some issues with compilation that need to be worked out in the
future. Specifically, attempting to compile too many instructions will
lead to an out-of-memory error. Also, dynamic compilation is slow. There's
no reason that we can't precompile all the instructions; this would
provide the speed of dynamic compilation without the slowdown of compiling
at runtime (and would provide an option for people who want to use this
with the JRE and not the JDK--the JDK is required for at-runtime
compilation). It would also help somewhat with the memory issues.