add InstructionProvider abstraction
The InstructionProvider interface represents the ability to provide
executable instructions given a numeric instruction value. The
InstructionCompiler has been refactored to be an InstructionProvider, and
the Cpu has been refactored to use one instead of hardcoding the
InstructionCompiler. This will make it easier to add more
InstructionProviders in the future (e.g. the precompiled instruction
provider).