add precompiled instruction provider
As each instruction (of ~50K) gets its own class, the precompiled
instruction provider does not make sense to include in the repository.
Instead, InstructionPrecompiler can be used to generate a precompiled
instruction provider. The code generation backend for the instructions is
the same as used for the dynamic instruction compiler. The performance is
on par with the dynamic instruction compiler, but when using precompiled
instructions it is not required to have the JDK installed (only the JRE).