make interrupts thread-safe
Previously it was necessary for hardware to wait until a safe time (e.g.
when cyclesElapsed() was called) before calling Cpu.interrupt(), as the
interrupt mechanism wasn't thread-safe. Replaced the thread queue with an
ArrayBlockingQueue, which is thread-safe. Also removed the code for the
Keyboard device that would wait for cyclesElapsed() to call interrupt(),
instead interrupting immediately.