simple performance improvement for terminal
Now, when the terminal blinks (i.e. any characters with the blink bit set
would reverse colors), if there are no characters with the blink bit set
active on the screen, the image is not marked as dirty.

This simple enhancement swapped drawGlyph() from using about 30% of
processor time in a CPU/terminal emulation loop to not even showing up on
the profiler's list. Of course this is only true if there are no blinked
characters on the screen, but blinking is rarely-enough used that this is
a worthwhile addition.