CPU: switch to ConcurrentLinkedQueue for interrupts
Previously, interrupts were stored in a LinkedBlockingQueue, which made
enforcing the maximum interrupt buffer size easy. However, this data
structure is kind of slow. Interrupts are now stored in a
ConcurrentLinkedQueue. An AtomicInteger is used to keep track of the
(approximate) number of interrupts in the queue. [The strict definition of
the counter is that it should contain less than or equal to the number of
entries in the queue data structure.]