Build: switch from Eclipse project to ant
Instead of using Eclipse to build, building is now done through ant. There
are two main entry points to the ant build file: "make", which will make
all of the normal source files and prepare a dynamic-compilation version
of the CPU; and "make-precompiled" which will do all of that as well as
generating a precompiled instruction jar for use.

Also:
  * Removed Eclipse-specific files that are not needed for ant building
  * Cleaned up the Terminal class so that it retrieves the default font
    image as a resource instead of hardcoding a filesystem path.