Core: call Device.attach() just before first step
Previously, Device.attach() was called in the Cpu's constructor. Calling
it later allows attached devices to use the initial memory state if they
wish to. Currently this is only used by the Debugger to break before the
first instruction is run. The debugger was also refactored to take
advantage of this functionality. (Previously, it had a special method
init() that needed to be called when attach() is now called.)