Debug: skip breakpoints on skipped instructions
Previously, even if the debugger frame was set to step over skipped instructions, any breakpoints on skipped instructions would still trigger a pause. The PC would point to the next instruction after skipping completed, potentially confusing the user as to why execution paused. Now, if the step-over-skipped option is set, any breakpoints on skipped instructions will be ignored. This also factors out the step-over-skipped functionality from the DebuggerFrame to the Debugger.