Assembler: remove uses of Either
Either<U, V> was a generic container that allowed a reference to one of two different classes. It was used by the assembler to represent values that were either known literal values, or a label that would need to be resolved later. But there's no reason to use such a generic class when a specific class is more expressive. Switched Either<U, V> to a new Assembly.Value class.